% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_template_files.R
\name{glm_template_files}
\alias{glm_template_files}
\title{Return the path to a template of current files for running GLM}
\usage{
glm_template_files()
}
\description{
This returns a path to a directory with example files for running GLM (configuration files and model forcing data).
}
\examples{
\dontrun{
glm_template_files()
}

}
\author{
Jorrit Mesman
}
\keyword{methods}
