% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/nml_template_path.R
\name{nml_template_path}
\alias{nml_template_path}
\title{Return the path to a current NML file template}
\usage{
nml_template_path()
}
\description{
The NML file can change with updated versions of GLM. This returns a path to a valid
NML template file matching the current GLM version.
}
\examples{
\dontrun{
file.edit(nml_template_path())
}
}
\author{
Luke Winslow, Jordan Read
}
\keyword{methods}

