% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_glm.R
\name{run_glm}
\alias{run_glm}
\title{Run the GLM model}
\usage{
run_glm(
  sim_folder = ".",
  nml_file = "glm4.nml",
  verbose = TRUE,
  system.args = character()
)
}
\arguments{
\item{sim_folder}{the directory where simulation files are contained}

\item{nml_file}{the configuration file that needs to be run}

\item{verbose}{Save output as character vector. Defaults to FALSE}
}
\description{
This runs the GLM model on the specific simulation stored in \code{sim_folder}.
The specified \code{sim_folder} must contain a valid .nml file.
}
\examples{
sim_folder <- system.file("extdata", package = "GLMr")
run_glm(sim_folder, nml_file = "glm4.nml")
}
\author{
Jorrit Mesman
}
\keyword{methods}
