% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_glm}
\alias{run_glm}
\title{run the GLM model}
\usage{
run_glm(sim_folder = ".", verbose = TRUE, args = character())
}
\arguments{
\item{sim_folder}{the directory where simulation files are contained}

\item{verbose}{should output of GLM be shown}

\item{args}{Optional arguments to pass to GLM executable}
}
\description{
This runs the GLM model on the specific simulation stored in \code{sim_folder}.
The specified \code{sim_folder} must contain a valid NML file.
}
\examples{
\dontrun{
library(glmtools)
sim_folder <- system.file('extdata/v2.0.0sim', package = 'GLMr')
run_glm(sim_folder)
out_file <- file.path(sim_folder,'output.nc')
nml_file <- system.file('extdata/v2.0.0sim', 'glm2.nml', package = 'GLMr')
field_file <- system.file('extdata', 'field_data.tsv', package = 'GLMr')
plot_temp(file = out_file, fig_path = 'test_temp.png')
summarize_sim(out_file, sim_outputs = c('temp','surface_height','wind'))

}
}
\author{
Jordan Read, Luke Winslow
}
\keyword{methods}

