% Generated by roxygen2 (4.0.2): do not edit by hand
\name{run_glm}
\alias{run_glm}
\title{run the GLM model}
\usage{
run_glm(sim_folder, verbose = TRUE)
}
\arguments{
\item{sim_folder}{the directory where simulation files are contained}

\item{verbose}{should output of GLM be shown}
}
\description{
This runs the GLM model on the specific simulation stored in \code{sim_folder}.
The specified \code{sim_folder} must contain a valid NML file.
}
\examples{
\dontrun{
library(glmtools)
sim_folder <- system.file('extdata/sim', package = 'GLMr')
run_glm(sim_folder)
out_file <- file.path(sim_folder,'output.nc')
nml_file <- system.file('extdata/sim', 'glm.nml', package = 'GLMr')
field_file <- system.file('extdata', 'field_data.tsv', package = 'GLMr')
plot_temp(file = out_file, fig_path = 'test_temp.png')
model_diagnostics(out_file, field_file, nml_file = nml_file,
   metrics = c('thermo.depth', 'schmidt.stability'),
   fig_path = FALSE)

}
}
\author{
Jordan Read, Luke Winslow
}
\keyword{methods}

